refactor(Leaflet): small code improvements
