+++ specifyweb/frontend/js_src/lib/components/QueryBuilder/FromMap.tsx
@@ -13,8 +13,7 @@ import type { QueryField } from './helpers';
 
 const emptyArray: RA<never> = [];
 const defaultPoint = [0, 10] as const;
-// REFACTOR: use this type where applicable
-export type Pair<T = number> = readonly [T, T];
+type Pair<T = number> = readonly [T, T];
 
 export function QueryFromMap({
   fields,
